requirements.txt
beautifulsoup4
lxml
python-telegram-bot
requests
gunicorn
